reorganize the rate unit tests this sorts them into categories so it is easier to work with
