+++ pynucastro/rates/tests/library_tests/README.md
@@ -1,0 +1,1 @@
+This holds tests that check the various Library member functions
