+++ pynucastro/rates/tests/library_tests/test_tabular_library.py
