+++ pynucastro/rates/tests/rate_tests/test_alternate_rates.py
@@ -1,4 +1,4 @@
-# unit tests for rates
+# unit tests for the alternate_rates module
 
 from pytest import approx
