+++ pynucastro/rates/tests/rate_tests/test_temperature_tabular.py
@@ -1,3 +1,5 @@
+# some tests of the TemperatureTabularRate interpolation
+
 import numpy as np
 import pytest
 from pytest import approx
