+++ pynucastro/rates/tests/weak_rate_tests/README.md
@@ -1,0 +1,2 @@
+This is intended to contain tests that check the value of the weak
+rate interpolation against the original source.
