+++ pynucastro/rates/tests/weak_rate_tests/conftest.py
@@ -1,0 +1,12 @@
+"""Fixtures used for rates testing."""
+
+# reuse fixtures from networks/tests
+# pylint: disable=unused-import
+from pynucastro.networks.tests.conftest import ffn_library  # noqa: F401
+from pynucastro.networks.tests.conftest import full_library  # noqa: F401
+from pynucastro.networks.tests.conftest import langanke_library  # noqa: F401
+from pynucastro.networks.tests.conftest import oda_library  # noqa: F401
+from pynucastro.networks.tests.conftest import reaclib_library  # noqa: F401
+from pynucastro.networks.tests.conftest import starlib_library  # noqa: F401
+from pynucastro.networks.tests.conftest import suzuki_library  # noqa: F401
+from pynucastro.networks.tests.conftest import tabular_library  # noqa: F401
