+++ pynucastro/rates/tests/weak_rate_tests/test_ffn_rates.py
@@ -1,3 +1,5 @@
+# test the FFN rates
+
 import pytest
 from pytest import approx
