+++ pynucastro/rates/tests/weak_rate_tests/test_langanke_rates.py
@@ -1,4 +1,4 @@
-# unit tests for rates
+# test the Langanke rates
 
 import sys
 import warnings
@@ -12,103 +12,8 @@ import pynucastro as pyna
 class TestTabularRates:
 
     @pytest.fixture(scope="class")
-    def rc_su(self, suzuki_library):
-        return pyna.RateCollection(libraries=[suzuki_library])
-
-    @pytest.fixture(scope="class")
     def rc_la(self, langanke_library):
         return pyna.RateCollection(libraries=[langanke_library])
-
-    def test_rate_values_suzuki(self, rc_su):
-
-        rho = 1.2e8
-        T = 1.5e9
-
-        comp_su = pyna.Composition(rc_su.get_nuclei())
-        comp_su.set_all(1)
-        comp_su.normalize()
-
-        # this is generated by doing
-        # rvals = rc_su.evaluate_rates(rho, T, comp_su)
-        # for r in sorted(rvals, key=lambda r: r.fname):
-        #     print(f"'{r.fname}': {rvals[r]},")
-
-        stored_rates_su = {
-            'Al23_to_Mg23_weaktab': 0.0018356969491904495,
-            'Al24_to_Mg24_weaktab': 0.0010224143375329186,
-            'Al25_to_Mg25_weaktab': 0.0004742839102314945,
-            'Al26_to_Mg26_weaktab': 8.208195176601902e-06,
-            'Al27_to_Mg27_weaktab': 1.3548922253387587e-12,
-            'Al28_to_Mg28_weaktab': 2.82771853373719e-14,
-            'Al28_to_Si28_weaktab': 6.552181957939603e-07,
-            'F17_to_O17_weaktab': 0.0004964162849900655,
-            'F18_to_O18_weaktab': 0.00014752795061742893,
-            'F19_to_O19_weaktab': 1.4589243454067581e-18,
-            'F20_to_Ne20_weaktab': 4.575609904051387e-05,
-            'F20_to_O20_weaktab': 1.413854271576516e-15,
-            'F21_to_Ne21_weaktab': 0.0001131073690508162,
-            'F21_to_O21_weaktab': 3.123684503103512e-30,
-            'F22_to_Ne22_weaktab': 0.0001215897114169894,
-            'F23_to_Ne23_weaktab': 0.00023974141956359604,
-            'Mg20_to_Na20_weaktab': 0.005114821722250049,
-            'Mg21_to_Na21_weaktab': 0.0027551736679093503,
-            'Mg22_to_Na22_weaktab': 0.0013109707173084527,
-            'Mg23_to_Na23_weaktab': 0.0004219912985023772,
-            'Mg24_to_Na24_weaktab': 1.6598635700843666e-22,
-            'Mg25_to_Na25_weaktab': 8.591632561336034e-16,
-            'Mg26_to_Na26_weaktab': 8.68429638614793e-33,
-            'Mg27_to_Al27_weaktab': 1.361348911272524e-08,
-            'Mg27_to_Na27_weaktab': 8.723314871252093e-32,
-            'Mg28_to_Al28_weaktab': 3.336720258691769e-11,
-            'Mg28_to_Na28_weaktab': 2.8134798652347447e-48,
-            'Na19_to_Ne19_weaktab': 0.002449472490421163,
-            'Na20_to_Ne20_weaktab': 0.002428040325297442,
-            'Na21_to_Ne21_weaktab': 0.00037796168976840364,
-            'Na22_to_Ne22_weaktab': 8.135557391945126e-07,
-            'Na23_to_Ne23_weaktab': 1.1720011968249513e-17,
-            'Na24_to_Mg24_weaktab': 4.406876111298294e-07,
-            'Na24_to_Ne24_weaktab': 3.756853129605931e-12,
-            'Na25_to_Mg25_weaktab': 4.479044974800567e-06,
-            'Na25_to_Ne25_weaktab': 3.72540094885677e-27,
-            'Na26_to_Mg26_weaktab': 0.001022078977999804,
-            'Na27_to_Mg27_weaktab': 0.0017911484289089184,
-            'Na28_to_Mg28_weaktab': 0.009005626791070679,
-            'Ne18_to_F18_weaktab': 0.0022422277113142404,
-            'Ne19_to_F19_weaktab': 0.0006992986372563014,
-            'Ne20_to_F20_weaktab': 9.47646722208835e-26,
-            'Ne21_to_F21_weaktab': 2.715931312189591e-21,
-            'Ne22_to_F22_weaktab': 9.580760966938022e-40,
-            'Ne23_to_F23_weaktab': 1.6479178669188904e-31,
-            'Ne23_to_Na23_weaktab': 9.3540132273414e-06,
-            'Ne24_to_Na24_weaktab': 1.0429568596930843e-07,
-            'Ne25_to_Na25_weaktab': 0.0007774750136737904,
-            'O17_to_F17_weaktab': 2.992303848236561e-23,
-            'O18_to_F18_weaktab': 6.178789007042831e-23,
-            'O19_to_F19_weaktab': 1.4762829483676722e-05,
-            'O20_to_F20_weaktab': 1.1564141869489023e-05,
-            'O21_to_F21_weaktab': 0.0001140275780950444,
-            'P27_to_Si27_weaktab': 0.0017552224872207786,
-            'P28_to_Si28_weaktab': 0.0011558955780814398,
-            'S28_to_P28_weaktab': 0.0025226735211831297,
-            'Si24_to_Al24_weaktab': 0.002836345548021039,
-            'Si25_to_Al25_weaktab': 0.0018713317001937818,
-            'Si26_to_Al26_weaktab': 0.0012411477482899727,
-            'Si27_to_Al27_weaktab': 0.0004907519533774708,
-            'Si28_to_Al28_weaktab': 3.1330669347239767e-18,
-        }
-
-        ys_su = comp_su.get_molar()
-
-        # the individual rate is
-        # r = Y(reactant) * table_value
-
-        for r in rc_su.get_rates():
-            assert r.source["Author"] == "T. Suzuki, H. Toki, K. Nomoto"
-            rr = ys_su[r.reactants[0]] * r.eval(T, rho=rho, comp=comp_su)
-            if r.fname in stored_rates_su:
-                assert rr == approx(stored_rates_su[r.fname], rel=1.e-6, abs=1.e-100), f"rate: {r} does not agree"
-            else:
-                warnings.warn(UserWarning(f"missing Suzuki tests for tabular rate {r}"))
 
     def test_rate_values_langanke(self, rc_la):
 
@@ -353,108 +258,3 @@ class TestTabularRates:
                 assert rr == approx(stored_rates_la[r.fname], rel=1.e-6, abs=1.e-100), f"rate: {r} does not agree"
             else:
                 warnings.warn(UserWarning(f"missing Langanke tests for tabular rate {r}"))
-
-    def test_nu_loss_values_suzuki(self, rc_su):
-
-        rho = 1.2e9
-        T = 1.5e9
-
-        comp_su = pyna.Composition(rc_su.get_nuclei())
-        comp_su.set_all(1)
-        comp_su.normalize()
-
-        # this is generated by doing
-        #
-        # for r in sorted(rc_su.rates, key=lambda r: r.fname):
-        #     nu_loss_rate = r.get_nu_loss(T, rho*comp_su.ye)
-        #     print(f"'{r.fname}': {nu_loss_rate},")
-
-        stored_nu_loss_su = {
-            'Al23_to_Mg23_weaktab': 0.0004372874399337381,
-            'Al24_to_Mg24_weaktab': 0.0002444261479368015,
-            'Al25_to_Mg25_weaktab': 0.00011500170073365321,
-            'Al26_to_Mg26_weaktab': 2.0956240501428204e-06,
-            'Al27_to_Mg27_weaktab': 3.255106279693442e-11,
-            'Al28_to_Mg28_weaktab': 1.49477663243628e-11,
-            'Al28_to_Si28_weaktab': 3.831060797866789e-14,
-            'F17_to_O17_weaktab': 7.899054617378834e-05,
-            'F18_to_O18_weaktab': 2.5282475946235158e-05,
-            'F19_to_O19_weaktab': 1.093403227437941e-14,
-            'F20_to_Ne20_weaktab': 2.1240115337229592e-08,
-            'F20_to_O20_weaktab': 2.057038605088727e-12,
-            'F21_to_Ne21_weaktab': 4.983256254206788e-08,
-            'F21_to_O21_weaktab': 3.478821084686598e-26,
-            'F22_to_Ne22_weaktab': 1.0792505314678724e-07,
-            'F23_to_Ne23_weaktab': 3.5106676050488424e-07,
-            'Mg20_to_Na20_weaktab': 0.0011319135934955277,
-            'Mg21_to_Na21_weaktab': 0.0006340336907068992,
-            'Mg22_to_Na22_weaktab': 0.0002767982218577626,
-            'Mg23_to_Na23_weaktab': 9.610438271090285e-05,
-            'Mg24_to_Na24_weaktab': 1.8757107528850667e-18,
-            'Mg25_to_Na25_weaktab': 5.385813495789488e-12,
-            'Mg26_to_Na26_weaktab': 1.042881660253991e-28,
-            'Mg27_to_Al27_weaktab': 2.6083434105025398e-15,
-            'Mg27_to_Na27_weaktab': 1.0899697988607703e-27,
-            'Mg28_to_Al28_weaktab': 5.426292791567733e-18,
-            'Mg28_to_Na28_weaktab': 6.450491997157852e-44,
-            'Na19_to_Ne19_weaktab': 0.000493293000357844,
-            'Na20_to_Ne20_weaktab': 0.0005899195182275463,
-            'Na21_to_Ne21_weaktab': 7.599282323751405e-05,
-            'Na22_to_Ne22_weaktab': 2.2006497168948355e-07,
-            'Na23_to_Ne23_weaktab': 1.2578502390300712e-13,
-            'Na24_to_Mg24_weaktab': 3.93370337581875e-10,
-            'Na24_to_Ne24_weaktab': 1.9814147087010928e-10,
-            'Na25_to_Mg25_weaktab': 2.6657308423828652e-11,
-            'Na25_to_Ne25_weaktab': 4.5166273695294323e-23,
-            'Na26_to_Mg26_weaktab': 5.681065910388462e-06,
-            'Na27_to_Mg27_weaktab': 7.802299251728987e-06,
-            'Na28_to_Mg28_weaktab': 9.465293245065728e-05,
-            'Ne18_to_F18_weaktab': 0.0003974021604949035,
-            'Ne19_to_F19_weaktab': 0.0001252832096143487,
-            'Ne20_to_F20_weaktab': 9.695265306659568e-22,
-            'Ne21_to_F21_weaktab': 2.923719748936441e-17,
-            'Ne22_to_F22_weaktab': 1.0542944818300268e-35,
-            'Ne23_to_F23_weaktab': 1.9118167540106322e-27,
-            'Ne23_to_Na23_weaktab': 4.3009430056097873e-10,
-            'Ne24_to_Na24_weaktab': 4.472165318443309e-17,
-            'Ne25_to_Na25_weaktab': 1.7531245231618863e-06,
-            'O17_to_F17_weaktab': 1.5965647193812905e-32,
-            'O18_to_F18_weaktab': 1.6695331795511284e-33,
-            'O19_to_F19_weaktab': 1.3591873287418124e-09,
-            'O20_to_F20_weaktab': 1.8668550656057127e-13,
-            'O21_to_F21_weaktab': 1.0304477129326074e-07,
-            'P27_to_Si27_weaktab': 0.0004940585002600768,
-            'P28_to_Si28_weaktab': 0.0003484280237167292,
-            'S28_to_P28_weaktab': 0.0007818005881530282,
-            'Si24_to_Al24_weaktab': 0.0007522480887757669,
-            'Si25_to_Al25_weaktab': 0.0005089386787967424,
-            'Si26_to_Al26_weaktab': 0.00031462787988451046,
-            'Si27_to_Al27_weaktab': 0.00013141083163038222,
-            'Si28_to_Al28_weaktab': 9.226341626912568e-15,
-        }
-
-        for r in rc_su.get_rates():
-            nu_loss = r.get_nu_loss(T, rho=rho, comp=comp_su)
-            if r.fname in stored_nu_loss_su:
-                assert nu_loss == approx(stored_nu_loss_su[r.fname], rel=1.e-6, abs=1.e-100), f"rate: {r} does not agree"
-            else:
-                warnings.warn(UserWarning(f"missing Suzuki tests for tabular nu loss rate {r}"))
-
-    def test_bounds(self, rc_su):
-
-        comp = pyna.Composition(rc_su.get_nuclei())
-        comp.set_all(1)
-        comp.normalize()
-
-        r = rc_su.get_rates()[0]
-
-        T = 1.e11
-        rho = 1.e7
-
-        with raises(ValueError):
-            r.eval(T, rho=rho, comp=comp)
-
-        rho = 1.e2
-
-        with raises(ValueError):
-            r.eval(T, rho=rho, comp=comp)
